빠르게 데이터 generation 하는 코드 작성
기초적인 prevalence, heritability 계산 코드도 작성

+++ Rng.h
@@ -38,7 +38,7 @@ void InitWELLRNG512a () {
     int j;
     __state_i = 0;
     for (j = 0; j < __R; j++)
-        __STATE[j] = time(NULL);
+        __STATE[j] = (unsigned int)time(NULL);
 }
 
 double Random() {
